+++ custom_components/aqua_temp/sensor.py
@@ -1,15 +1,18 @@
 import logging
-import sys
 
 from homeassistant.components.sensor import SensorDeviceClass, SensorEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
-from homeassistant.core import HomeAssistant
-from homeassistant.helpers.entity_platform import AddEntitiesCallback
+from homeassistant.core import HomeAssistant, callback
+from homeassistant.helpers.dispatcher import async_dispatcher_connect
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.util import slugify
 
-from .common.consts import ALL_ENTITIES, DOMAIN
+from .common.consts import DOMAIN, SIGNAL_AQUA_TEMP_DEVICE_NEW
+from .common.device_discovery import (
+    async_handle_discovered_device,
+    find_entity_description,
+)
 from .common.entity_descriptions import AquaTempSensorEntityDescription
 from .managers.aqua_temp_coordinator import AquaTempCoordinator
 
@@ -17,34 +20,35 @@ _LOGGER = logging.getLogger(__name__)
 
 
 async def async_setup_entry(
-    hass: HomeAssistant, entry: ConfigEntry, async_add_entities: AddEntitiesCallback
+    hass: HomeAssistant, entry: ConfigEntry, async_add_entities
 ):
-    """Set up the sensor platform."""
-    try:
+    def _create(
+        device_code: str, entity_description_key: str, coordinator: AquaTempCoordinator
+    ) -> AquaTempSensorEntity:
+        entity_description = find_entity_description(
+            entity_description_key, Platform.SENSOR
+        )
+
+        entity = AquaTempSensorEntity(device_code, entity_description, coordinator)
+
+        return entity
+
+    @callback
+    def _async_device_new(device_code):
         coordinator = hass.data[DOMAIN][entry.entry_id]
-        entities = []
-        entity_descriptions = []
 
-        for entity_description in ALL_ENTITIES:
-            if entity_description.platform == Platform.SENSOR:
-                entity_descriptions.append(entity_description)
+        async_handle_discovered_device(
+            device_code,
+            coordinator,
+            Platform.SENSOR,
+            _create,
+            async_add_entities,
+        )
 
-        for device_code in coordinator.api_data:
-            for entity_description in entity_descriptions:
-                entity = AquaTempSensorEntity(
-                    device_code, entity_description, coordinator
-                )
-
-                entities.append(entity)
-
-        _LOGGER.debug(f"Setting up sensor entities: {entities}")
-
-        async_add_entities(entities, True)
-    except Exception as ex:
-        exc_type, exc_obj, tb = sys.exc_info()
-        line_number = tb.tb_lineno
-
-        _LOGGER.error(f"Failed to initialize sensor, Error: {ex}, Line: {line_number}")
+    """Set up the binary sensor platform."""
+    entry.async_on_unload(
+        async_dispatcher_connect(hass, SIGNAL_AQUA_TEMP_DEVICE_NEW, _async_device_new)
+    )
 
 
 class AquaTempSensorEntity(CoordinatorEntity, SensorEntity):
@@ -59,20 +63,20 @@ class AquaTempSensorEntity(CoordinatorEntity, SensorEntity):
         super().__init__(coordinator)
 
         self._device_code = device_code
-        self._api_data = self.coordinator.api_data[self._device_code]
-        self._config_data = self.coordinator.config_data
+        self._config_data = coordinator.config_data
 
         device_info = coordinator.get_device(device_code)
         device_name = device_info.get("name")
+        device_data = coordinator.get_device_data(self._device_code)
 
         entity_name = f"{device_name} {entity_description.name}"
 
         slugify_name = slugify(entity_name)
 
-        device_id = self._api_data.get("device_id")
+        device_id = device_data.get("device_id")
         unique_id = slugify(f"{entity_description.platform}_{slugify_name}_{device_id}")
 
-        self.entity_description: AquaTempSensorEntityDescription = entity_description
+        self.entity_description = entity_description
 
         self._attr_device_info = device_info
         self._attr_name = entity_name
@@ -84,14 +88,18 @@ class AquaTempSensorEntity(CoordinatorEntity, SensorEntity):
 
         if entity_description.device_class == SensorDeviceClass.TEMPERATURE:
             self._attr_native_unit_of_measurement = (
-                self.coordinator.get_temperature_unit(device_code)
+                self._local_coordinator.get_temperature_unit(device_code)
             )
+
+    @property
+    def _local_coordinator(self) -> AquaTempCoordinator:
+        return self.coordinator
 
     def _handle_coordinator_update(self) -> None:
         """Fetch new state data for the sensor."""
-        state: float | int | str | None = self._api_data.get(
-            self.entity_description.key
-        )
+        device_data = self._local_coordinator.get_device_data(self._device_code)
+
+        state = device_data.get(self.entity_description.key)
 
         if isinstance(state, str):
             state = float(state)