+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v3.0.4
+
+- Add device discovery for new devices without need to reload integration
+- Fix integration reload issue
+- Entities code clean up
+- API stabilization after timeout
+- Fix target temperature parameter mapping (Heat, Cool and Auto)
+- Fix parameter mapping for P01 (Water pump mode)
+
 ## v3.0.3
 
 - Fix setting HVAC mode